# apps/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Couture Dynamic Pro — API métier (Phase 4)
  version: "1.0"
  description: >
    Contrat des endpoints REST versionnés (collections, catalogues, pages,
    produits, variantes, panier, commandes). Les endpoints /api/auth et
    /api/admin/users (Phase 3) et les paiements (Phase 7) ne sont pas
    couverts par ce document.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cdp_session
...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
      description: Requis pour les opérations sensibles (ex. création de commande).
paths:
...
  /panier/lignes:
    post:
      summary: Ajouter une ligne au panier (fusion de quantité si la variante y est déjà)
      description: Rôle autorisé — CLIENT uniquement. RG6 — disponibilité vérifiée côté serveur.
      responses:
        "200": { description: OK }
        "409": { description: Stock insuffisant (INSUFFICIENT_STOCK) }
  /panier/lignes/{ligneId}:
    patch:
      summary: Modifier la quantité d'une ligne
      description: Rôle autorisé — CLIENT uniquement (sur son propre panier).
      parameters:
        - name: ligneId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
        "404": { description: Ligne non trouvée (hors périmètre) }
    delete:
      summary: Retirer une ligne du panier
      description: Rôle autorisé — CLIENT uniquement (sur son propre panier).
      parameters:
        - name: ligneId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
  /commandes:
    get:
      summary: Lister les commandes (RG2 — périmètre par rôle)
      description: >
        CLIENT : ses commandes uniquement. LOGISTICS : uniquement
        PREPARING/SHIPPED (son périmètre). ADMIN : toutes, filtrable par
        ?status=.
      responses:
        "200": { description: OK }
    post:
      summary: Passer commande (checkout) — snapshot immuable (RG7), stock décrémenté atomiquement
      description: >
        Rôle autorisé — CLIENT uniquement. Idempotente (Idempotency-Key
        requis) : un rejeu avec la même clé renvoie la réponse déjà produite
        sans recréer de commande. Aucun paiement initié ici (Phase 7).
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "201": { description: Créée }
        "400": { description: Panier vide ou adresse de livraison manquante }
        "409": { description: Stock insuffisant, conflit de devise, ou rejeu en cours }
  /commandes/{id}:
    get:
      summary: Détail d'une commande (404 hors périmètre du rôle, jamais 403 — anti-énumération)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
        "404": { description: Non trouvée / hors périmètre }
  /commandes/{id}/statut:
    patch:
      summary: Transition de statut (RG9)
      description: >
        Rôles autorisés — ADMIN (RECEIVED→PREPARING/CANCELLED,
        PREPARING→SHIPPED/CANCELLED, SHIPPED→DELIVERED), LOGISTICS
        (PREPARING→SHIPPED, SHIPPED→DELIVERED uniquement). FAILED n'est pas
        atteignable ici : déclenché par un échec de paiement (Phase 7).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
        "409": { description: Transition non autorisée (INVALID_TRANSITION) }